+++ main-service/src/main/java/ru/practicum/repository/LocationRepository.java
@@ -3,9 +3,10 @@ package ru.practicum.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.model.Location;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
 
-    Optional<Location> findByLatAndLon(Float lat, Float lon);
+    Optional<Location> findByLatAndLon(BigDecimal lat, BigDecimal lon);
 }
